Quick and dirty port to .NET 9

+++ ClassPutty.cs
@@ -57,7 +57,7 @@ namespace GitForce
                         {
                             // Send a notification to Pageant to close
                             IntPtr winHandle = NativeMethods.FindWindow("PageantSysTray", null);
-                            NativeMethods.SendMessage(winHandle, NativeMethods.WM_COMMAND, NativeMethods.WM_CLOSE, 0);
+                            NativeMethods.SendMessage(winHandle, NativeMethods.WM_COMMAND, (nint)NativeMethods.WM_CLOSE, 0);
                         }
                         //File.Delete(_pathPageant);
                     }
